Optimized polar discriminator as only imaginary part was returned.
Original 21ns/sample
Optimized 9ns/sample 

Optimized frequency shift leveraging symmetrical shift/rotation needs only 4 multiplications and 4 additions instead of 8 multiplications and 4 additions. 
Independent shifting 47ns/sample
Symmetric shifting  23ns/sample

Profiler data frequency shift opt:
Old asymmetric
 %   cumulative   self              self     total           
 time   seconds   seconds    calls  ms/call  ms/call  name    
   4.30     46.46     2.47 111542272     0.00     0.00  shift_freq_minus400
  4.13     48.83     2.37 111542272     0.00     0.00  shift_freq_250
New symmetric:
  %   cumulative   self              self     total           
 time   seconds   seconds    calls  ms/call  ms/call  name    
 5.54     40.17     2.57 111542272     0.00     0.00  shift_freq_plus_minus325

Polar discriminator opt:
Original
  %   cumulative   self              self     total           
 time   seconds   seconds    calls  ms/call  ms/call  name    
   2.20     50.09     1.27 55771136     0.00     0.00  polar_discriminator_t1_c1
  1.80     51.13     1.04 55771136     0.00     0.00  polar_discriminator_s1
Optimized
  1.19     43.71     0.55 55771136     0.00     0.00  polar_discriminator_t1_c1
  0.99     44.17     0.46 55771136     0.00     0.00  polar_discriminator_s1