@xaelsouth 2021-02-15
Hello alalons, thanks for suggested optimizations!

I could not merge your patches as is. I had to merge them by copying line by line from your patches complementing with my comments, even further optimizations and some beautification after all that ;-). 

I would appreciate your code review and test for simultaneous receiving of C1/T1 ans S1 at 868.625M. Could you share to me a new file with both datagram types included?

Xael